Reorganize `Storage.LedgerDB` internal modules
