+++ ouroboros-consensus-cardano-tools/src/Cardano/Tools/DBAnalyser/Run.hs
@@ -30,9 +30,8 @@ import           Ouroboros.Consensus.Util.ResourceRegistry
 import qualified Ouroboros.Consensus.Storage.ChainDB as ChainDB
 import           Ouroboros.Consensus.Storage.ChainDB.Impl.Args (fromChainDbArgs)
 import qualified Ouroboros.Consensus.Storage.ImmutableDB as ImmutableDB
-import           Ouroboros.Consensus.Storage.LedgerDB.DiskPolicy
-                     (SnapshotInterval (..), defaultDiskPolicy)
-import           Ouroboros.Consensus.Storage.LedgerDB.OnDisk (readSnapshot)
+import           Ouroboros.Consensus.Storage.LedgerDB (SnapshotInterval (..),
+                     defaultDiskPolicy, readSnapshot)
 import qualified Ouroboros.Consensus.Storage.VolatileDB as VolatileDB
 
 import           Cardano.Tools.DBAnalyser.Analysis
