+++ ouroboros-consensus/changelog.d/20230127_150735_javier.sagredo_reorganize_consensus_storage_ledgerdb.md
@@ -1,0 +1,6 @@
+### Non-Breaking
+
+- Reorganized `Ouroboros.Consensus.Storage.LedgerDB.*` modules. Old modules have
+  a deprecation warning for downstream users but otherwise they still export the
+  same functionality.
+
