+++ ouroboros-consensus/src/Ouroboros/Consensus/Storage/LedgerDB/LedgerDB.hs
@@ -1,0 +1,137 @@
+{-# LANGUAGE DeriveAnyClass        #-}
+{-# LANGUAGE DeriveGeneric         #-}
+{-# LANGUAGE FlexibleContexts      #-}
+{-# LANGUAGE FlexibleInstances     #-}
+{-# LANGUAGE MultiParamTypeClasses #-}
+{-# LANGUAGE StandaloneDeriving    #-}
+{-# LANGUAGE TypeFamilies          #-}
+{-# LANGUAGE UndecidableInstances  #-}
+
+module Ouroboros.Consensus.Storage.LedgerDB.LedgerDB (
+    -- * LedgerDB
+    Checkpoint (..)
+  , LedgerDB (..)
+  , LedgerDB'
+  , LedgerDbCfg (..)
+  , configLedgerDb
+  ) where
+
+import           GHC.Generics (Generic)
+
+import           NoThunks.Class (NoThunks)
+
+import           Ouroboros.Network.AnchoredSeq (Anchorable (..),
+                     AnchoredSeq (..))
+import qualified Ouroboros.Network.AnchoredSeq as AS
+
+import           Ouroboros.Consensus.Block
+import           Ouroboros.Consensus.Config
+import           Ouroboros.Consensus.Ledger.Abstract
+import           Ouroboros.Consensus.Ledger.Extended (ExtLedgerCfg (..),
+                     ExtLedgerState)
+import           Ouroboros.Consensus.Protocol.Abstract (ConsensusProtocol)
+
+{-------------------------------------------------------------------------------
+  LedgerDB
+-------------------------------------------------------------------------------}
+
+-- | Internal state of the ledger DB
+--
+-- The ledger DB looks like
+--
+-- > anchor |> snapshots <| current
+--
+-- where @anchor@ records the oldest known snapshot and @current@ the most
+-- recent. The anchor is the oldest point we can roll back to.
+--
+-- We take a snapshot after each block is applied and keep in memory a window
+-- of the last @k@ snapshots. We have verified empirically (#1936) that the
+-- overhead of keeping @k@ snapshots in memory is small, i.e., about 5%
+-- compared to keeping a snapshot every 100 blocks. This is thanks to sharing
+-- between consecutive snapshots.
+--
+-- As an example, suppose we have @k = 6@. The ledger DB grows as illustrated
+-- below, where we indicate the anchor number of blocks, the stored snapshots,
+-- and the current ledger.
+--
+-- > anchor |> #   [ snapshots ]                   <| tip
+-- > ---------------------------------------------------------------------------
+-- > G      |> (0) [ ]                             <| G
+-- > G      |> (1) [ L1]                           <| L1
+-- > G      |> (2) [ L1,  L2]                      <| L2
+-- > G      |> (3) [ L1,  L2,  L3]                 <| L3
+-- > G      |> (4) [ L1,  L2,  L3,  L4]            <| L4
+-- > G      |> (5) [ L1,  L2,  L3,  L4,  L5]       <| L5
+-- > G      |> (6) [ L1,  L2,  L3,  L4,  L5,  L6]  <| L6
+-- > L1     |> (6) [ L2,  L3,  L4,  L5,  L6,  L7]  <| L7
+-- > L2     |> (6) [ L3,  L4,  L5,  L6,  L7,  L8]  <| L8
+-- > L3     |> (6) [ L4,  L5,  L6,  L7,  L8,  L9]  <| L9   (*)
+-- > L4     |> (6) [ L5,  L6,  L7,  L8,  L9,  L10] <| L10
+-- > L5     |> (6) [*L6,  L7,  L8,  L9,  L10, L11] <| L11
+-- > L6     |> (6) [ L7,  L8,  L9,  L10, L11, L12] <| L12
+-- > L7     |> (6) [ L8,  L9,  L10, L12, L12, L13] <| L13
+-- > L8     |> (6) [ L9,  L10, L12, L12, L13, L14] <| L14
+--
+-- The ledger DB must guarantee that at all times we are able to roll back @k@
+-- blocks. For example, if we are on line (*), and roll back 6 blocks, we get
+--
+-- > L3 |> []
+newtype LedgerDB l = LedgerDB {
+      -- | Ledger states
+      ledgerDbCheckpoints :: AnchoredSeq
+                               (WithOrigin SlotNo)
+                               (Checkpoint l)
+                               (Checkpoint l)
+    }
+  deriving (Generic)
+
+type LedgerDB' blk = LedgerDB (ExtLedgerState blk)
+
+deriving instance Show     l => Show     (LedgerDB l)
+deriving instance Eq       l => Eq       (LedgerDB l)
+deriving instance NoThunks l => NoThunks (LedgerDB l)
+
+type instance HeaderHash (LedgerDB l) = HeaderHash l
+
+instance IsLedger l => GetTip (LedgerDB l) where
+  getTip = castPoint
+         . getTip
+         . either unCheckpoint unCheckpoint
+         . AS.head
+         . ledgerDbCheckpoints
+
+-- | Internal newtype wrapper around a ledger state @l@ so that we can define a
+-- non-blanket 'Anchorable' instance.
+newtype Checkpoint l = Checkpoint {
+      unCheckpoint :: l
+    }
+  deriving (Generic)
+
+deriving instance Show     l => Show     (Checkpoint l)
+deriving instance Eq       l => Eq       (Checkpoint l)
+deriving instance NoThunks l => NoThunks (Checkpoint l)
+
+instance GetTip l => Anchorable (WithOrigin SlotNo) (Checkpoint l) (Checkpoint l) where
+  asAnchor = id
+  getAnchorMeasure _ = getTipSlot . unCheckpoint
+
+{-------------------------------------------------------------------------------
+  LedgerDB Config
+-------------------------------------------------------------------------------}
+
+data LedgerDbCfg l = LedgerDbCfg {
+      ledgerDbCfgSecParam :: !SecurityParam
+    , ledgerDbCfg         :: !(LedgerCfg l)
+    }
+  deriving (Generic)
+
+deriving instance NoThunks (LedgerCfg l) => NoThunks (LedgerDbCfg l)
+
+configLedgerDb ::
+     ConsensusProtocol (BlockProtocol blk)
+  => TopLevelConfig blk
+  -> LedgerDbCfg (ExtLedgerState blk)
+configLedgerDb cfg = LedgerDbCfg {
+      ledgerDbCfgSecParam = configSecurityParam cfg
+    , ledgerDbCfg         = ExtLedgerCfg cfg
+    }
