+++ ouroboros-consensus/src/Ouroboros/Consensus/Storage/LedgerDB/OnDisk.hs
@@ -1,566 +1,5 @@
-{-# LANGUAGE BangPatterns        #-}
-{-# LANGUAGE DeriveGeneric       #-}
-{-# LANGUAGE FlexibleContexts    #-}
-{-# LANGUAGE KindSignatures      #-}
-{-# LANGUAGE LambdaCase          #-}
-{-# LANGUAGE MultiWayIf          #-}
-{-# LANGUAGE NamedFieldPuns      #-}
-{-# LANGUAGE RankNTypes          #-}
-{-# LANGUAGE RecordWildCards     #-}
-{-# LANGUAGE ScopedTypeVariables #-}
-{-# LANGUAGE TupleSections       #-}
-{-# LANGUAGE TypeApplications    #-}
-
-module Ouroboros.Consensus.Storage.LedgerDB.OnDisk (
-    -- * Opening the database
-    InitFailure (..)
-  , InitLog (..)
-  , initLedgerDB
-    -- ** Instantiate in-memory to @blk@
-  , AnnLedgerError'
-  , LedgerDB'
-    -- ** Abstraction over the stream API
-  , NextBlock (..)
-  , StreamAPI (..)
-    -- * Read from disk
-  , readSnapshot
-    -- * Write to disk
-  , takeSnapshot
-  , trimSnapshots
-  , writeSnapshot
-    -- * Low-level API (primarily exposed for testing)
-  , deleteSnapshot
-  , snapshotToFileName
-  , snapshotToPath
-    -- ** opaque
-  , DiskSnapshot (..)
-    -- * Trace events
-  , ReplayGoal (..)
-  , ReplayStart (..)
-  , TraceEvent (..)
-  , TraceReplayEvent (..)
-  , decorateReplayTracerWithGoal
+module Ouroboros.Consensus.Storage.LedgerDB.OnDisk {-# DEPRECATED "Use Ouroboros.Consensus.Storage.LedgerDB instead" #-}
+  ( module Ouroboros.Consensus.Storage.LedgerDB
   ) where
 
-import qualified Codec.CBOR.Write as CBOR
-import           Codec.Serialise.Decoding (Decoder)
-import           Codec.Serialise.Encoding (Encoding)
-import           Control.Monad.Except
-import           Control.Tracer
-import qualified Data.List as List
-import           Data.Maybe (isJust, mapMaybe)
-import           Data.Ord (Down (..))
-import           Data.Set (Set)
-import qualified Data.Set as Set
-import           Data.Word
-import           GHC.Generics (Generic)
-import           GHC.Stack
-import           Text.Read (readMaybe)
-
-import           Ouroboros.Network.Block (Point (Point))
-
-import           Ouroboros.Consensus.Block
-import           Ouroboros.Consensus.HeaderValidation
-                     (HeaderState (headerStateTip), annTipPoint)
-import           Ouroboros.Consensus.Ledger.Abstract
-import           Ouroboros.Consensus.Ledger.Extended
-import           Ouroboros.Consensus.Ledger.Inspect
-import           Ouroboros.Consensus.Ledger.SupportsProtocol
-import           Ouroboros.Consensus.Util.CBOR (ReadIncrementalErr,
-                     readIncremental)
-import           Ouroboros.Consensus.Util.IOLike
-
-import           Ouroboros.Consensus.Storage.FS.API
-import           Ouroboros.Consensus.Storage.FS.API.Types
-
-import           Ouroboros.Consensus.Storage.LedgerDB.DiskPolicy
-import           Ouroboros.Consensus.Storage.LedgerDB.InMemory
-
-{-------------------------------------------------------------------------------
-  Instantiate the in-memory DB to @blk@
--------------------------------------------------------------------------------}
-
-type LedgerDB'       blk = LedgerDB       (ExtLedgerState blk)
-type AnnLedgerError' blk = AnnLedgerError (ExtLedgerState blk) blk
-
-{-------------------------------------------------------------------------------
-  Abstraction over the streaming API provided by the Chain DB
--------------------------------------------------------------------------------}
-
--- | Next block returned during streaming
-data NextBlock blk = NoMoreBlocks | NextBlock blk
-
--- | Stream blocks from the immutable DB
---
--- When we initialize the ledger DB, we try to find a snapshot close to the
--- tip of the immutable DB, and then stream blocks from the immutable DB to its
--- tip to bring the ledger up to date with the tip of the immutable DB.
---
--- In CPS form to enable the use of 'withXYZ' style iterator init functions.
-data StreamAPI m blk = StreamAPI {
-      -- | Start streaming after the specified block
-      streamAfter :: forall a. HasCallStack
-        => Point blk
-        -- Reference to the block corresponding to the snapshot we found
-        -- (or 'GenesisPoint' if we didn't find any)
-
-        -> (Either (RealPoint blk) (m (NextBlock blk)) -> m a)
-        -- Get the next block (by value)
-        --
-        -- Should be @Left pt@ if the snapshot we found is more recent than the
-        -- tip of the immutable DB. Since we only store snapshots to disk for
-        -- blocks in the immutable DB, this can only happen if the immutable DB
-        -- got truncated due to disk corruption. The returned @pt@ is a
-        -- 'RealPoint', not a 'Point', since it must always be possible to
-        -- stream after genesis.
-        -> m a
-    }
-
--- | Stream all blocks
-streamAll ::
-     forall m blk e a. (Monad m, HasCallStack)
-  => StreamAPI m blk
-  -> Point blk             -- ^ Starting point for streaming
-  -> (RealPoint blk -> e)  -- ^ Error when tip not found
-  -> a                     -- ^ Starting point when tip /is/ found
-  -> (blk -> a -> m a)     -- ^ Update function for each block
-  -> ExceptT e m a
-streamAll StreamAPI{..} tip notFound e f = ExceptT $
-    streamAfter tip $ \case
-      Left tip' -> return $ Left (notFound tip')
-
-      Right getNext -> do
-        let go :: a -> m a
-            go a = do mNext <- getNext
-                      case mNext of
-                        NoMoreBlocks -> return a
-                        NextBlock b  -> go =<< f b a
-        Right <$> go e
-
-{-------------------------------------------------------------------------------
-  Initialize the DB
--------------------------------------------------------------------------------}
-
--- | Initialization log
---
--- The initialization log records which snapshots from disk were considered,
--- in which order, and why some snapshots were rejected. It is primarily useful
--- for monitoring purposes.
-data InitLog blk =
-    -- | Defaulted to initialization from genesis
-    --
-    -- NOTE: Unless the blockchain is near genesis, we should see this /only/
-    -- if data corrupted occurred.
-    InitFromGenesis
-
-    -- | Used a snapshot corresponding to the specified tip
-  | InitFromSnapshot DiskSnapshot (RealPoint blk)
-
-    -- | Initialization skipped a snapshot
-    --
-    -- We record the reason why it was skipped.
-    --
-    -- NOTE: We should /only/ see this if data corrupted occurred.
-  | InitFailure DiskSnapshot (InitFailure blk) (InitLog blk)
-  deriving (Show, Eq, Generic)
-
--- | Initialize the ledger DB from the most recent snapshot on disk
---
--- If no such snapshot can be found, use the genesis ledger DB. Returns the
--- initialized DB as well as the block reference corresponding to the snapshot
--- we found on disk (the latter primarily for testing/monitoring purposes).
---
--- We do /not/ catch any exceptions thrown during streaming; should any be
--- thrown, it is the responsibility of the 'ChainDB' to catch these
--- and trigger (further) validation. We only discard snapshots if
---
--- * We cannot deserialise them, or
--- * they are /ahead/ of the chain
---
--- It is possible that the Ledger DB will not be able to roll back @k@ blocks
--- after initialization if the chain has been truncated (data corruption).
---
--- We do /not/ attempt to use multiple ledger states from disk to construct the
--- ledger DB. Instead we load only a /single/ ledger state from disk, and
--- /compute/ all subsequent ones. This is important, because the ledger states
--- obtained in this way will (hopefully) share much of their memory footprint
--- with their predecessors.
-initLedgerDB ::
-     forall m blk. (
-         IOLike m
-       , LedgerSupportsProtocol blk
-       , InspectLedger blk
-       , HasCallStack
-       )
-  => Tracer m (ReplayGoal blk -> TraceReplayEvent blk)
-  -> Tracer m (TraceEvent blk)
-  -> SomeHasFS m
-  -> (forall s. Decoder s (ExtLedgerState blk))
-  -> (forall s. Decoder s (HeaderHash blk))
-  -> LedgerDbCfg (ExtLedgerState blk)
-  -> m (ExtLedgerState blk) -- ^ Genesis ledger state
-  -> StreamAPI m blk
-  -> m (InitLog blk, LedgerDB' blk, Word64)
-initLedgerDB replayTracer
-             tracer
-             hasFS
-             decLedger
-             decHash
-             cfg
-             getGenesisLedger
-             streamAPI = do
-    snapshots <- listSnapshots hasFS
-    tryNewestFirst id snapshots
-  where
-    tryNewestFirst :: (InitLog blk -> InitLog blk)
-                   -> [DiskSnapshot]
-                   -> m (InitLog blk, LedgerDB' blk, Word64)
-    tryNewestFirst acc [] = do
-        -- We're out of snapshots. Start at genesis
-        traceWith replayTracer ReplayFromGenesis
-        initDb <- ledgerDbWithAnchor <$> getGenesisLedger
-        let replayTracer' = decorateReplayTracerWithStart (Point Origin) replayTracer
-        ml     <- runExceptT $ initStartingWith replayTracer' cfg streamAPI initDb
-        case ml of
-          Left _  -> error "invariant violation: invalid current chain"
-          Right (l, replayed) -> return (acc InitFromGenesis, l, replayed)
-    tryNewestFirst acc (s:ss) = do
-        -- If we fail to use this snapshot, delete it and try an older one
-        ml <- runExceptT $ initFromSnapshot
-                             replayTracer
-                             hasFS
-                             decLedger
-                             decHash
-                             cfg
-                             streamAPI
-                             s
-        case ml of
-          Left err -> do
-            when (diskSnapshotIsTemporary s) $
-              -- We don't delete permanent snapshots, even if we couldn't parse
-              -- them
-              deleteSnapshot hasFS s
-            traceWith tracer $ InvalidSnapshot s err
-            tryNewestFirst (acc . InitFailure s err) ss
-          Right (r, l, replayed) ->
-            return (acc (InitFromSnapshot s r), l, replayed)
-
-{-------------------------------------------------------------------------------
-  Internal: initialize using the given snapshot
--------------------------------------------------------------------------------}
-
-data InitFailure blk =
-    -- | We failed to deserialise the snapshot
-    --
-    -- This can happen due to data corruption in the ledger DB.
-    InitFailureRead ReadIncrementalErr
-
-    -- | This snapshot is too recent (ahead of the tip of the chain)
-  | InitFailureTooRecent (RealPoint blk)
-
-    -- | This snapshot was of the ledger state at genesis, even though we never
-    -- take snapshots at genesis, so this is unexpected.
-  | InitFailureGenesis
-  deriving (Show, Eq, Generic)
-
--- | Attempt to initialize the ledger DB from the given snapshot
---
--- If the chain DB or ledger layer reports an error, the whole thing is aborted
--- and an error is returned. This should not throw any errors itself (ignoring
--- unexpected exceptions such as asynchronous exceptions, of course).
-initFromSnapshot ::
-     forall m blk. (
-         IOLike m
-       , LedgerSupportsProtocol blk
-       , InspectLedger blk
-       , HasCallStack
-       )
-  => Tracer m (ReplayGoal blk -> TraceReplayEvent blk)
-  -> SomeHasFS m
-  -> (forall s. Decoder s (ExtLedgerState blk))
-  -> (forall s. Decoder s (HeaderHash blk))
-  -> LedgerDbCfg (ExtLedgerState blk)
-  -> StreamAPI m blk
-  -> DiskSnapshot
-  -> ExceptT (InitFailure blk) m (RealPoint blk, LedgerDB' blk, Word64)
-initFromSnapshot tracer hasFS decLedger decHash cfg streamAPI ss = do
-    initSS <- withExceptT InitFailureRead $
-                readSnapshot hasFS decLedger decHash ss
-    let initialPoint = withOrigin (Point Origin) annTipPoint $ headerStateTip $ headerState $ initSS
-    case pointToWithOriginRealPoint (castPoint (getTip initSS)) of
-      Origin        -> throwError InitFailureGenesis
-      NotOrigin tip -> do
-        lift $ traceWith tracer $ ReplayFromSnapshot ss tip (ReplayStart initialPoint)
-        let tracer' = decorateReplayTracerWithStart initialPoint tracer
-        (initDB, replayed) <-
-          initStartingWith
-            tracer'
-            cfg
-            streamAPI
-            (ledgerDbWithAnchor initSS)
-        return (tip, initDB, replayed)
-
--- | Attempt to initialize the ledger DB starting from the given ledger DB
-initStartingWith ::
-     forall m blk. (
-         Monad m
-       , LedgerSupportsProtocol blk
-       , InspectLedger blk
-       , HasCallStack
-       )
-  => Tracer m (ReplayStart blk -> ReplayGoal blk -> TraceReplayEvent blk)
-  -> LedgerDbCfg (ExtLedgerState blk)
-  -> StreamAPI m blk
-  -> LedgerDB' blk
-  -> ExceptT (InitFailure blk) m (LedgerDB' blk, Word64)
-initStartingWith tracer cfg streamAPI initDb = do
-    streamAll streamAPI (castPoint (ledgerDbTip initDb))
-      InitFailureTooRecent
-      (initDb, 0)
-      push
-  where
-    push :: blk -> (LedgerDB' blk, Word64) -> m (LedgerDB' blk, Word64)
-    push blk !(!db, !replayed) = do
-        !db' <- ledgerDbPush cfg (ReapplyVal blk) db
-
-        let replayed' :: Word64
-            !replayed' = replayed + 1
-
-            events :: [LedgerEvent blk]
-            events = inspectLedger
-                       (getExtLedgerCfg (ledgerDbCfg cfg))
-                       (ledgerState (ledgerDbCurrent db))
-                       (ledgerState (ledgerDbCurrent db'))
-
-        traceWith tracer (ReplayedBlock (blockRealPoint blk) events)
-        return (db', replayed')
-
-{-------------------------------------------------------------------------------
-  Write to disk
--------------------------------------------------------------------------------}
-
--- | Take a snapshot of the /oldest ledger state/ in the ledger DB
---
--- We write the /oldest/ ledger state to disk because the intention is to only
--- write ledger states to disk that we know to be immutable. Primarily for
--- testing purposes, 'takeSnapshot' returns the block reference corresponding
--- to the snapshot that we wrote.
---
--- If a snapshot with the same number already exists on disk or if the tip is at
--- genesis, no snapshot is taken.
---
--- Note that an EBB can have the same slot number and thus snapshot number as
--- the block after it. This doesn't matter. The one block difference in the
--- ledger state doesn't warrant an additional snapshot. The number in the name
--- of the snapshot is only indicative, we don't rely on it being correct.
---
--- NOTE: This is a lower-level API that takes a snapshot independent from
--- whether this snapshot corresponds to a state that is more than @k@ back.
---
--- TODO: Should we delete the file if an error occurs during writing?
-takeSnapshot ::
-     forall m blk. (MonadThrow m, IsLedger (LedgerState blk))
-  => Tracer m (TraceEvent blk)
-  -> SomeHasFS m
-  -> (ExtLedgerState blk -> Encoding)
-  -> LedgerDB' blk -> m (Maybe (DiskSnapshot, RealPoint blk))
-takeSnapshot tracer hasFS encLedger db =
-    case pointToWithOriginRealPoint (castPoint (getTip oldest)) of
-      Origin ->
-        return Nothing
-      NotOrigin tip -> do
-        let number   = unSlotNo (realPointSlot tip)
-            snapshot = DiskSnapshot number Nothing
-        snapshots <- listSnapshots hasFS
-        if List.any ((== number) . dsNumber) snapshots then
-          return Nothing
-        else do
-          writeSnapshot hasFS encLedger snapshot oldest
-          traceWith tracer $ TookSnapshot snapshot tip
-          return $ Just (snapshot, tip)
-  where
-    oldest :: ExtLedgerState blk
-    oldest = ledgerDbAnchor db
-
--- | Trim the number of on disk snapshots so that at most 'onDiskNumSnapshots'
--- snapshots are stored on disk. The oldest snapshots are deleted.
---
--- The deleted snapshots are returned.
-trimSnapshots ::
-     Monad m
-  => Tracer m (TraceEvent r)
-  -> SomeHasFS m
-  -> DiskPolicy
-  -> m [DiskSnapshot]
-trimSnapshots tracer hasFS DiskPolicy{..} = do
-    -- We only trim temporary snapshots
-    snapshots <- filter diskSnapshotIsTemporary <$> listSnapshots hasFS
-    -- The snapshot are most recent first, so we can simply drop from the
-    -- front to get the snapshots that are "too" old.
-    forM (drop (fromIntegral onDiskNumSnapshots) snapshots) $ \snapshot -> do
-      deleteSnapshot hasFS snapshot
-      traceWith tracer $ DeletedSnapshot snapshot
-      return snapshot
-
-{-------------------------------------------------------------------------------
-  Internal: reading from disk
--------------------------------------------------------------------------------}
-
-data DiskSnapshot = DiskSnapshot {
-      -- | Snapshots are numbered. We will try the snapshots with the highest
-      -- number first.
-      --
-      -- When creating a snapshot, we use the slot number of the ledger state it
-      -- corresponds to as the snapshot number. This gives an indication of how
-      -- recent the snapshot is.
-      --
-      -- Note that the snapshot names are only indicative, we don't rely on the
-      -- snapshot number matching the slot number of the corresponding ledger
-      -- state. We only use the snapshots numbers to determine the order in
-      -- which we try them.
-      dsNumber :: Word64
-
-      -- | Snapshots can optionally have a suffix, separated by the snapshot
-      -- number with an underscore, e.g., @4492799_last_Byron@. This suffix acts
-      -- as metadata for the operator of the node. Snapshots with a suffix will
-      -- /not be trimmed/.
-    , dsSuffix :: Maybe String
-    }
-  deriving (Show, Eq, Ord, Generic)
-
--- | Named snapshot are permanent, they will never be deleted when trimming.
-diskSnapshotIsPermanent :: DiskSnapshot -> Bool
-diskSnapshotIsPermanent = isJust . dsSuffix
-
--- | The snapshots that are periodically created are temporary, they will be
--- deleted when trimming
-diskSnapshotIsTemporary :: DiskSnapshot -> Bool
-diskSnapshotIsTemporary = not . diskSnapshotIsPermanent
-
--- | Read snapshot from disk
-readSnapshot ::
-     forall m blk. IOLike m
-  => SomeHasFS m
-  -> (forall s. Decoder s (ExtLedgerState blk))
-  -> (forall s. Decoder s (HeaderHash blk))
-  -> DiskSnapshot
-  -> ExceptT ReadIncrementalErr m (ExtLedgerState blk)
-readSnapshot hasFS decLedger decHash =
-      ExceptT
-    . readIncremental hasFS decoder
-    . snapshotToPath
-  where
-    decoder :: Decoder s (ExtLedgerState blk)
-    decoder = decodeSnapshotBackwardsCompatible (Proxy @blk) decLedger decHash
-
--- | Write snapshot to disk
-writeSnapshot ::
-     forall m blk. MonadThrow m
-  => SomeHasFS m
-  -> (ExtLedgerState blk -> Encoding)
-  -> DiskSnapshot
-  -> ExtLedgerState blk -> m ()
-writeSnapshot (SomeHasFS hasFS) encLedger ss cs = do
-    withFile hasFS (snapshotToPath ss) (WriteMode MustBeNew) $ \h ->
-      void $ hPut hasFS h $ CBOR.toBuilder (encode cs)
-  where
-    encode :: ExtLedgerState blk -> Encoding
-    encode = encodeSnapshot encLedger
-
--- | Delete snapshot from disk
-deleteSnapshot :: HasCallStack => SomeHasFS m -> DiskSnapshot -> m ()
-deleteSnapshot (SomeHasFS HasFS{..}) = removeFile . snapshotToPath
-
--- | List on-disk snapshots, highest number first.
-listSnapshots :: Monad m => SomeHasFS m -> m [DiskSnapshot]
-listSnapshots (SomeHasFS HasFS{..}) =
-    aux <$> listDirectory (mkFsPath [])
-  where
-    aux :: Set String -> [DiskSnapshot]
-    aux = List.sortOn (Down . dsNumber) . mapMaybe snapshotFromPath . Set.toList
-
-snapshotToFileName :: DiskSnapshot -> String
-snapshotToFileName DiskSnapshot { dsNumber, dsSuffix } =
-    show dsNumber <> suffix
-  where
-    suffix = case dsSuffix of
-      Nothing -> ""
-      Just s  -> "_" <> s
-
-snapshotToPath :: DiskSnapshot -> FsPath
-snapshotToPath = mkFsPath . (:[]) . snapshotToFileName
-
-snapshotFromPath :: String -> Maybe DiskSnapshot
-snapshotFromPath fileName = do
-    number <- readMaybe prefix
-    return $ DiskSnapshot number suffix'
-  where
-    (prefix, suffix) = break (== '_') fileName
-
-    suffix' :: Maybe String
-    suffix' = case suffix of
-      ""      -> Nothing
-      _ : str -> Just str
-
-{-------------------------------------------------------------------------------
-  Trace events
--------------------------------------------------------------------------------}
-
--- | Add the tip of the Immutable DB to the trace event
---
--- Between the tip of the immutable DB and the point of the starting block,
--- the node could (if it so desired) easily compute a "percentage complete".
-decorateReplayTracerWithGoal
-  :: Point blk -- ^ Tip of the ImmutableDB
-  -> Tracer m (TraceReplayEvent blk)
-  -> Tracer m (ReplayGoal blk -> TraceReplayEvent blk)
-decorateReplayTracerWithGoal immTip = contramap ($ (ReplayGoal immTip))
-
--- | Add the block at which a replay started.
---
--- This allows to compute a "percentage complete" when tracing the events.
-decorateReplayTracerWithStart
-  :: Point blk -- ^ Starting point of the replay
-  -> Tracer m (ReplayGoal blk -> TraceReplayEvent blk)
-  -> Tracer m (ReplayStart blk -> ReplayGoal blk -> TraceReplayEvent blk)
-decorateReplayTracerWithStart start = contramap ($ (ReplayStart start))
-
-data TraceEvent blk
-  = InvalidSnapshot DiskSnapshot (InitFailure blk)
-    -- ^ An on disk snapshot was skipped because it was invalid.
-  | TookSnapshot DiskSnapshot (RealPoint blk)
-    -- ^ A snapshot was written to disk.
-  | DeletedSnapshot DiskSnapshot
-    -- ^ An old or invalid on-disk snapshot was deleted
-  deriving (Generic, Eq, Show)
-
--- | Which point the replay started from
-newtype ReplayStart blk = ReplayStart (Point blk) deriving (Eq, Show)
-
--- | Which point the replay is expected to end at
-newtype ReplayGoal blk = ReplayGoal (Point blk) deriving (Eq, Show)
-
--- | Events traced while replaying blocks against the ledger to bring it up to
--- date w.r.t. the tip of the ImmutableDB during initialisation. As this
--- process takes a while, we trace events to inform higher layers of our
--- progress.
-data TraceReplayEvent blk
-  = -- | There were no LedgerDB snapshots on disk, so we're replaying all blocks
-    -- starting from Genesis against the initial ledger.
-    ReplayFromGenesis
-        (ReplayGoal blk)  -- ^ the block at the tip of the ImmutableDB
-    -- | There was a LedgerDB snapshot on disk corresponding to the given tip.
-    -- We're replaying more recent blocks against it.
-  | ReplayFromSnapshot
-        DiskSnapshot
-        (RealPoint blk)
-        (ReplayStart blk) -- ^ the block at which this replay started
-        (ReplayGoal blk)  -- ^ the block at the tip of the ImmutableDB
-  -- | We replayed the given block (reference) on the genesis snapshot during
-  -- the initialisation of the LedgerDB. Used during ImmutableDB replay.
-  | ReplayedBlock
-        (RealPoint blk)   -- ^ the block being replayed
-        [LedgerEvent blk]
-        (ReplayStart blk) -- ^ the block at which this replay started
-        (ReplayGoal blk)  -- ^ the block at the tip of the ImmutableDB
-  deriving (Generic, Eq, Show)
+import Ouroboros.Consensus.Storage.LedgerDB
